refactor: put fw rules table in a component
